Refactor and remove non essential files
