+++ activity/kafka/descriptor.json
