--- activity/kafkapub/activity_test.go
@@ -1,16 +1,0 @@
-package kafkapub
-
-import (
-	"testing"
-
-	"github.com/project-flogo/core/activity"
-	"github.com/stretchr/testify/assert"
-)
-
-func TestRegister(t *testing.T) {
-
-	ref := activity.GetRef(&KafkaPubActivity{})
-	act := activity.Get(ref)
-
-	assert.NotNil(t, act)
-}
