+++ activity/log/go.mod
@@ -3,4 +3,6 @@ module github.com/project-flogo/contrib/activity/log
 require (
 	github.com/project-flogo/core v0.9.0-beta.2
 	github.com/stretchr/testify v1.3.0
+	go.uber.org/atomic v1.3.2 // indirect
+	go.uber.org/multierr v1.1.0 // indirect
 )
