+++ trigger/kafka/descriptor.json
