+++ trigger/kafka/go.mod
@@ -1,4 +1,4 @@
-module github.com/project-flogo/contrib/trigger/kafkasub
+module github.com/project-flogo/contrib/trigger/kafka
 
 require (
 	github.com/Shopify/sarama v1.22.0
