+++ trigger/kafka/metadata.go
@@ -1,4 +1,4 @@
-package kafkasub
+package kafka
 
 import (
 	"github.com/project-flogo/core/data/coerce"
