+++ trigger/rest/go.mod
@@ -4,6 +4,4 @@ require (
 	github.com/julienschmidt/httprouter v1.2.0
 	github.com/project-flogo/core v0.9.0-beta.2
 	github.com/stretchr/testify v1.3.0
-	go.uber.org/atomic v1.3.2 // indirect
-	go.uber.org/multierr v1.1.0 // indirect
 )
